Implements additional tests

+++ api/app/Http/Controllers/BaseController.php
@@ -34,32 +34,6 @@ abstract class BaseController extends Controller
     protected $transformer = 'App\Http\Transformers\BaseTransformer';
 
     /**
-     * Transform a collection for response.
-     *
-     * @param  Collection $collection
-     * @return array
-     */
-    protected function collection(Collection $collection)
-    {
-        $transformer = App::make('App\Services\Transformer');
-
-        return $transformer->collection($collection, new $this->transformer);
-    }
-
-    /**
-     * Transform an item for response.
-     *
-     * @param  Model  $item
-     * @return array
-     */
-    protected function item(Model $item)
-    {
-        $transformer = App::make('App\Services\Transformer');
-
-        return $transformer->item($item, new $this->transformer);
-    }
-
-    /**
      * Get all entities.
      *
      * @return Response
@@ -137,7 +111,11 @@ abstract class BaseController extends Controller
      */
     public function patchOne($id, Request $request)
     {
-        return $this->repository->update($id, $request->all());
+        if (!$this->validator->with(array_add($request->all(), 'entity_id', $id))->patch()->passes()) {
+            return $this->validator->errors();
+        }
+
+        return (string) $this->repository->update($id, $request->all());
     }
 
     /**
@@ -148,7 +126,12 @@ abstract class BaseController extends Controller
      */
     public function patchMany(Request $request)
     {
-        return $this->repository->updateMany($request->all());
+        foreach ($request->data as $entity) {
+            if (!$this->validator->with($entity)->patch()->passes()) {
+                return $this->validator->errors();
+            }
+        }
+        return (string) $this->repository->updateMany($request->data);
     }
 
     /**
@@ -159,6 +142,10 @@ abstract class BaseController extends Controller
      */
     public function deleteOne($id)
     {
+        if (!$this->validator->delete()->with(['entity_id' => $id])->passes()) {
+            return $this->validator->errors();
+        }
+
         return $this->repository->delete($id);
     }
 
@@ -170,7 +157,38 @@ abstract class BaseController extends Controller
      */
     public function deleteMany(Request $request)
     {
-        return $this->repository->deleteMany($request->all());
+        foreach ($request->data as $id) {
+            if (!$this->validator->delete()->with(['entity_id' => $id])->passes()) {
+                return $this->validator->errors();
+            }
+        }
+        return $this->repository->deleteMany($request->data);
+    }
+
+    /**
+     * Transform an item for response.
+     *
+     * @param  Model  $item
+     * @return array
+     */
+    protected function item(Model $item)
+    {
+        $transformer = App::make('App\Services\Transformer');
+
+        return $transformer->item($item, new $this->transformer);
+    }
+
+    /**
+     * Transform a collection for response.
+     *
+     * @param  Collection $collection
+     * @return array
+     */
+    protected function collection(Collection $collection)
+    {
+        $transformer = App::make('App\Services\Transformer');
+
+        return $transformer->collection($collection, new $this->transformer);
     }
 
     public static function renderException($request, \Exception $e, $debug = false){
